fix crash and error with delete, fix validator

+++ resources/views/users/index.blade.php
@@ -19,25 +19,26 @@
 
 <script>
     $('#user-table').on('click', '.btn-delete', function (e) { 
-    var token = '{{ csrf_token() }}';
-    e.preventDefault();
-    $.ajaxSetup({
-        headers: {
-            'X-CSRF-TOKEN': token,
-            'X-Requested-With': 'XMLHttpRequest',
+        if(confirm("Are you sure want to delete?")){
+            var token = '{{ csrf_token() }}';
+            e.preventDefault();
+            $.ajaxSetup({
+                headers: {
+                    'X-CSRF-TOKEN': token,
+                    'X-Requested-With': 'XMLHttpRequest',
+                }
+            });
+            var url = $(this).data('id-variable');
+            // confirm then
+            $.ajax({
+                url: url,
+                type: 'DELETE',
+                dataType: 'json',
+                data: {method: '_DELETE', submit: true}
+            }).always(function (data) {
+                $('#user-table').DataTable().draw(false);
+            });
         }
     });
-    var url = $(this).data('id-variable');
-    // confirm then
-    $.ajax({
-        url: url,
-        type: 'DELETE',
-        dataType: 'json',
-        data: {method: '_DELETE', submit: true}
-    }).always(function (data) {
-        $('#user-table').DataTable().draw(false);
-    });
-});
-
 </script>
 @endpush
